hours.el: Keep track of days when calculating invoice
It's not perfect yet: If you have 2 hours entries on the same day it will
count it as 2 days. It needs to roll those up into a single day.

Along the same lines, something that starts at 12am or 2am should be rolled
in with the previous day and not the next day.